implement cb feedback + merge fixes
